// src/types/Orioles.ts
// src/types/Oriole.ts
export interface Oriole {
  id: number;
  bird_name: string;
  scientific_name: string;
  picture_01: string;
  picture_02: string;
  picture_03: string;
  range_picture: string;
  urlslug: string;
  wiki: string;
  length_cm: string;
  weight_gram: string;
  synopsis: string;
};